combined tests into one file
